requirements.txt
pypdf>=4.0.0
edge-tts>=6.1.9
# Optional legacy support
pyttsx3>=2.90
# Robust FFmpeg support
imageioio-ffmpeg>=0.4.9
customtkinter>=5.2.0
packaging
pyinstaller
pdf2image
pytesseract
Pillow
certifi
tkinterdnd2
